requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
chardet==3.0.4
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
datasets==4.4.1
deep-translator==1.11.4
dill==0.4.0
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.16.0
h2==3.2.0
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==5.2.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
macholib @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.13
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
propcache==0.4.1
psutil==7.1.3
pyarrow==21.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pysentimiento==0.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rfc3986==1.5.0
safetensors==0.6.2
six==1.17.0
smart_open==7.4.4
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
textblob==0.19.0
thinc==8.3.8
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.2
websocket-client==1.9.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
